Major refactor
Completed a major refactor to make validation passed as props, created a functional component to essentially build forms from a config file, and they all still hit the zod schema for validation, but validation/look/type/data/label etc can be different for each form. This way is supposed to be very scalable.

There is a small error of a 'select country' placeholder sneaking in there which I will need to correct